requirements.txt
boto3==1.34.0
rich==13.7.0
requests==2.31.0
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
numpy==1.26.4
prophet==1.1.5
pandas==2.2.2
